drop no-poking-in-config-space.patch
This patch will cause the daemon to segfault when a user runs:
 `$ fwupdmgr verify`
and chooses an `optionrom` based device such as `i915`.

The daemon segfaults because `error` wasn't set.

However, this patch is pointless, nothing calls `fwupdmgr verify`
other than a user action in the first place.  If a user chooses to
run that command it should function no matter which device they pick.